+++ src/routers/petitions/petitions.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter
 
 from model.Schema.petition import CreatePetition, ViewPetition
 
-from controller.petitions_controller import counting_petition, new_petition
+from controller.petitions import counting_petition, new_petition
 
 
 petitions = APIRouter()
